[Refactor] YearMonth, InfoType을 TempModel에서 분리
