+++ BudgetBuddies/BudgetBuddies/Sources/Screen/Calendar/Calendar/Helpers/AlertManager.swift
@@ -22,6 +22,7 @@ class AlertManager {
   ///   - title: 알림의 제목을 나타내는 문자열입니다.
   ///   - message: 알림의 내용이나 메세지를 나타내는 문자열입니다. 메세지가 필요 없는 경우 'nil'을 사용할 수 있습니다.
   ///   - needsCancelButton: 'true'일 경우 취소 버튼이 추가되고, 'false'일 경우 추가되지 않습니다.
+    ///   
   public static func showAlert(
     on vc: UIViewController, title: String, message: String?, needsCancelButton: Bool
   ) {
