+++ BudgetBuddies/BudgetBuddies/Sources/Screen/Calendar/Calendar/Models/InfoType.swift
@@ -1,0 +1,13 @@
+//
+//  InfoType.swift
+//  BudgetBuddies
+//
+//  Created by 김승원 on 8/9/24.
+//
+
+import Foundation
+
+enum InfoType {
+    case discount
+    case support
+}
